refactor: main page api 호출 수정
