+++ src/app/(home)/page.tsx
@@ -5,13 +5,20 @@ import PostsLink from './_components/PostsLink';
 import TodayQna from './_components/TodayQna';
 import TopButton from '../../components/TopButton';
 
-const HomePage = () => {
+const HomePage = async () => {
+  const response = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/main-page`, {
+    next: {
+      revalidate: 3600
+    }
+  });
+  const data = await response.json();
+  console.log(data);
   return (
     <div className="flex flex-col justify-center items-center relative">
       <div className="xl:w-[1200px] md:w-[800px] sm:w-96 flex flex-col gap-20">
         <PostsLink />
-        <BestForum />
-        <TodayQna />
+        <BestForum forumList={data?.forum_posts} />
+        <TodayQna todayQna={data?.qna_posts} />
         <LandingPage />
         <MainPageTag />
         <TopButton />
